+++ src/chara.rs
@@ -4,7 +4,7 @@
 
 use std::{collections::{HashMap, HashSet}, cmp::{min, max}};
 use rand::{rngs::ThreadRng, Rng};
-use crate::{util::*, board::Board, gen::Zobrist, engine::search};
+use crate::{util::*, board::Board, gen::Zobrist, engine::search, weights::Weights};
 
 // TODO: use f16 instead of f32 to cache approx ~40 mil./2 GiB more positions?
 // TODO: use i16 instead of floats for faster add/sub operations?
@@ -61,9 +61,8 @@ impl Chara {
 	// It's a little messy, but any evals are easily modifiable.
 	// No need in modifying something twice to be applied for the different coloir, neither there's a need to write something twice in eval()
     pub fn init(board: &Board, random_range: f32, aggressiveness: f32) -> Chara {
-		let zobrist = Zobrist::default();
+		let mut w = Weights::default();
 
-		// let piece_wmult: f32					= 1.5;	-- Result were somewhat better
 		let piece_wmult: f32					= 1.2;
 		let piece_square_related_wmult: f32		= 0.75 / (aggressiveness * 0.5  + 0.5 );	// really slight balance-out
 		let mobility_wmult: f32					= 1.0  * (aggressiveness * 0.5  + 0.5 );
@@ -72,268 +71,60 @@ impl Chara {
 		let pawn_structure_wmult: f32			= 0.5  * (aggressiveness * 0.5  + 0.5 );
 		let minor_piece_pos_wmult: f32			= 1.0  * (aggressiveness * 0.5  + 0.5 );
 
-		/* 2-staged static piece positional weights
-			- they will be tuned, for now I need to focus no search */
-		/* Inspiration for a current implementation was taken from:
-			- Simplified Evaluation Function by Tomasz Michniewski
-			- PeSTO's Evaluation Function by Tom Kerrigan, Pawel Koziol, Ronald Friederich (The OG!!)
-			Note: some modifications are ill-intended for fun sake, copying is not advised :D */
-		// TODO: tune them, add more dynamic types of weights
-		// TODO (optional): store weights in file maybe to possibly autotune them later
-		let pieces_weights_square_related = [
-			// opening or middlegame
-			[
-                // pawns
-				[
-					 0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,
-					 0.96,  1.21,  0.7 ,  0.95,  0.71,  1.2 ,  0.6 ,  0.0 ,
-					-0.05,  0.05,  0.25,  0.3 ,  0.3 ,  0.5 ,  0.25,  0.1 ,
-					-0.15,  0.05,  0.05,  0.2 ,  0.25,  0.1 ,  0.1 , -0.15,
-					-0.2 , -0.02,  0.0 ,  0.15,  0.2 ,  0.0 ,  0.1 , -0.2 ,
-					-0.25, -0.03, -0.1 ,  0.0 ,  0.0 , -0.3 ,  0.3 ,  0.0 ,
-					-0.3 , -0.01, -0.15, -0.15, -0.15,  0.3 ,  0.3 , -0.15,
-					 0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0
-				],
-                // knights
-				[
-					-1.63, -0.9 , -0.3 , -0.49,  0.01, -0.8 , -0.2 , -1.03,
-					-0.72, -0.4 ,  0.71,  0.35,  0.2 ,  0.6 ,  0.01, -0.05,
-					-0.4 ,  0.6 ,  0.4 ,  0.65,  0.85,  1.3 ,  0.72,  0.4 ,
-					-0.05,  0.2 ,  0.2 ,  0.54,  0.4 ,  0.7 ,  0.22,  0.2 ,
-					-0.1 ,  0.0 ,  0.15,  0.21,  0.21,  0.2 ,  0.2 , -0.02,
-					-0.3 , -0.08,  0.15,  0.1 ,  0.15,  0.2 ,  0.25, -0.15,
-					-0.3 , -0.5 , -0.15,  0.01,  0.01,  0.0 , -0.1 , -0.2 ,
-					-1.1 , -0.5 , -0.6 , -0.1 , -0.1 , -0.02, -0.5 , -0.25
-				],
-                // bishops
-				[
-					-0.3 , -0.65, -0.5 , -0.35, -0.25, -0.3 , -0.6 , -0.1 ,
-					-0.25,  0.2 , -0.15, -0.05,  0.3 ,  0.6 ,  0.2 , -0.4 ,
-					-0.1 ,  0.3 ,  0.4 ,  0.4 ,  0.35,  0.5 ,  0.37,  0.0 ,
-					 0.0 ,  0.15,  0.11,  0.5 ,  0.37,  0.37,  0.15,  0.0 ,
-					 0.0 ,  0.1 ,  0.3 ,  0.2 ,  0.2 ,  0.12,  0.1 ,  0.0 ,
-					 0.02,  0.15,  0.15,  0.05,  0.0 ,  0.0 ,  0.05,  0.06,
-					 0.0 ,  0.15,  0.05,  0.0 ,  0.1 ,  0.15,  0.3 ,  0.0 ,
-					-0.3 ,  0.0 , -0.15, -0.15, -0.15, -0.1 , -0.4 , -0.25
-				],
-                // rooks
-				[
-					 0.3 ,  0.4 ,  0.3 ,  0.5 ,  0.6 ,  0.1 ,  0.3 ,  0.4 ,
-					 0.3 ,  0.3 ,  0.6 ,  0.6 ,  0.8 ,  0.6 ,  0.3 ,  0.4 ,
-					-0.05,  0.15,  0.2 ,  0.35,  0.2 ,  0.45,  0.5 ,  0.13,
-					-0.25, -0.1 ,  0.05,  0.2 ,  0.2 ,  0.25, -0.05, -0.2 ,
-					-0.3 , -0.2 , -0.05,  0.0 ,  0.05, -0.05,  0.05, -0.25,
-					-0.46, -0.2 , -0.15, -0.1 ,  0.05,  0.0 , -0.05, -0.3 ,
-					-0.4 , -0.12, -0.13, -0.01,  0.0 ,  0.1 , -0.07, -0.5 ,
-					-0.15, -0.09,  0.03,  0.2 ,  0.3 ,  0.07, -0.39, -0.2
-				],
-                // queens
-				[
-					-0.29,  0.0 ,  0.3 ,  0.11,  0.6 ,  0.44,  0.43,  0.45,
-					-0.22, -0.4 , -0.01,  0.01, -0.15,  0.57,  0.25,  0.5 ,
-					-0.15, -0.15,  0.05,  0.05,  0.3 ,  0.55,  0.47,  0.56,
-					-0.25, -0.25, -0.15, -0.15,  0.0 ,  0.15,  0.0 ,  0.0 ,
-					-0.05, -0.25, -0.1 , -0.1 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,
-					-0.1 ,  0.05, -0.1 , -0.01, -0.04,  0.0 ,  0.11,  0.01,
-					-0.3 ,  0.0 ,  0.11,  0.02,  0.07,  0.05, -0.01,  0.0 ,
-					-0.1 , -0.2 , -0.02,  0.1 , -0.15, -0.25, -0.3 , -0.5
-				],
-                // king
-				[
-					-0.4 , -0.4 , -0.4 , -0.4 , -0.1 , -0.4 , -0.4 , -0.35,
-					-0.35,  0.4 , -0.4 , -0.4 , -0.4 , -0.4 ,  0.35, -0.3 ,
-					-0.3 , -0.35, -0.4 , -0.4 , -0.4 , -0.35, -0.3 , -0.2 ,
-					-0.25, -0.3 , -0.35, -0.4 , -0.4 , -0.3 , -0.2 , -0.3 ,
-					-0.2 , -0.25, -0.3 , -0.35, -0.35, -0.2 , -0.25, -0.2 ,
-					-0.15, -0.2 , -0.25, -0.3 , -0.05, -0.3 , -0.2 , -0.1 ,
-					 0.0 ,  0.01, -0.05, -0.25, -0.1 , -0.1 ,  0.03,  0.02,
-					-0.2 ,  0.3 ,  0.2 , -0.2 ,  0.1 , -0.1 ,  0.3 ,  0.2
-				]
-            ],
-			// endgame
-			[
-                // pawns
-				[
-					 0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,
-					 1.95,  1.8 ,  1.6 ,  1.4 ,  1.5 ,  1.3 ,  1.65,  1.9 ,
-					 0.95,  1.0 ,  0.9 ,  0.7 ,  0.55,  0.5 ,  0.8 ,  0.85,
-					 0.3 ,  0.2 ,  0.19,  0.05,  0.03,  0.04,  0.2 ,  0.2 ,
-					 0.15, -0.1 , -0.04, -0.05, -0.05, -0.1 ,  0.0 , -0.01,
-					 0.0 ,  0.05, -0.05,  0.0 ,  0.0 , -0.05,  0.0 , -0.1 ,
-					-0.2 , -0.2 , -0.2 , -0.3 , -0.3 , -0.01, -0.2 , -0.3 ,
-					 0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0
-				],
-                // knights
-				[
-					-0.6 , -0.45, -0.15, -0.3 , -0.3 , -0.3 , -0.45, -0.99,
-					-0.25, -0.2 , -0.1 ,  0.0 ,  0.0 , -0.1 , -0.2 , -0.45,
-					-0.2 , -0.1 ,  0.05,  0.1 ,  0.1 ,  0.01, -0.1 , -0.2 ,
-					-0.2 ,  0.05,  0.15,  0.2 ,  0.2 ,  0.1 ,  0.05, -0.2 ,
-					-0.2 ,  0.0 ,  0.1 ,  0.2 ,  0.2 ,  0.1 ,  0.0 , -0.2 ,
-					-0.2 , -0.1 ,  0.0 ,  0.1 ,  0.1 ,  0.0 , -0.2 , -0.2 ,
-					-0.45, -0.2 , -0.2 , -0.05, -0.05, -0.2 , -0.2 , -0.45,
-					-0.7 , -0.45, -0.3 , -0.3 , -0.3 , -0.3 , -0.45, -0.7
-				],
-				// bishops
-				[
-					-0.3 , -0.08, -0.05, -0.05, -0.05, -0.05, -0.08, -0.3 ,
-					-0.08, -0.01,  0.0 , -0.04, -0.03,  0.0 , -0.01, -0.08,
-					 0.05,  0.0 ,  0.05,  0.0 ,  0.0 ,  0.06,  0.0 ,  0.05,
-					 0.0 ,  0.06,  0.1 ,  0.05,  0.06,  0.1 ,  0.03,  0.0 ,
-					 0.0 ,  0.05,  0.1 ,  0.11,  0.05,  0.1 ,  0.0 ,  0.0 ,
-					-0.05,  0.0 ,  0.06,  0.04,  0.04,  0.01,  0.0 , -0.05,
-					-0.15, -0.1 ,  0.0 ,  0.0 ,  0.0 ,  0.0 , -0.1 , -0.15,
-					-0.25, -0.15, -0.2 , -0.1 , -0.1 , -0.2 , -0.15, -0.25
-				],
-				// rooks (it kinda makes sense)
-				[
-					 0.1 ,  0.1 ,  0.1 ,  0.1 ,  0.1 ,  0.1 ,  0.0 ,  0.0 ,
-					 0.1 ,  0.1 ,  0.1 ,  0.1 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,
-					 0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,
-					 0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,
-					 0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,
-					 0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,
-					 0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,
-					 0.0 ,  0.04,  0.04,  0.0 ,  0.0 ,  0.0 ,  0.04, -0.11
-				],
-                // queens
-				[
-					-0.01,  0.14,  0.14,  0.16,  0.16,  0.14,  0.14,  0.01,
-					-0.1 ,  0.2 ,  0.3 ,  0.45,  0.6 ,  0.25,  0.3 ,  0.01,
-					-0.2 ,  0.05,  0.05,  0.5 ,  0.45,  0.25,  0.2 ,  0.01,
-					-0.1 ,  0.2 ,  0.2 ,  0.45,  0.55,  0.4 ,  0.4 ,  0.05,
-					-0.14,  0.2 ,  0.2 ,  0.45,  0.35,  0.3 ,  0.3 ,  0.05,
-					-0.16, -0.05,  0.1 ,  0.04,  0.05,  0.1 ,  0.05,  0.01,
-					-0.22, -0.18, -0.18, -0.16, -0.06, -0.18, -0.18, -0.22,
-					-0.33, -0.22, -0.22, -0.4 , -0.1 , -0.3 , -0.22, -0.4
-				],
-                // king
-				[
-					-0.5 , -0.3 , -0.2 , -0.2 , -0.1 , -0.05, -0.05, -0.3 ,
-					-0.1 ,  0.15,  0.15,  0.16,  0.16,  0.3 ,  0.2 ,  0.05,
-					 0.1 ,  0.2 ,  0.2 ,  0.2 ,  0.2 ,  0.46,  0.45,  0.1 ,
-					-0.05,  0.2 ,  0.21,  0.31,  0.31,  0.3 ,  0.25,  0.0 ,
-					-0.2 ,  0.0 ,  0.2 ,  0.30,  0.30,  0.2 ,  0.05, -0.1 ,
-					-0.2 ,  0.0 ,  0.1 ,  0.2 ,  0.2 ,  0.1 ,  0.05, -0.1 ,
-					-0.25, -0.1 ,  0.05,  0.1 ,  0.1 ,  0.05, -0.05, -0.15,
-					-0.5 , -0.3 , -0.3 , -0.3 , -0.3 , -0.3 , -0.3 , -0.5
-				]
-			]
-		];
-		// 0 for a king seems sus but we don't really care bc we don't calc its capture
-		let pieces_weights_const = [
-			[ 0.85,  3.3 ,  3.0 ,  4.0 ,  9.0 ,  0.0 ],
-			[ 0.9 ,  3.0 ,  3.1 ,  5.0 ,  9.0 ,  0.0 ]
-		];
 		// transform (flip for white, negative for black) and apply coefficients
 		let mut pieces_weights = [[[0.0; 64]; 12]; 2];
 		for i in 0..2 {
 			for j in 0..6 {
 				for k in 0..64 {
-					pieces_weights[i][ j << 1     ][k] =  pieces_weights_square_related[i][j][flip(k)] * piece_square_related_wmult + pieces_weights_const[i][j] * piece_wmult;
-					pieces_weights[i][(j << 1) | 1][k] = -pieces_weights_square_related[i][j][k      ] * piece_square_related_wmult - pieces_weights_const[i][j] * piece_wmult;
+					pieces_weights[i][ j << 1     ][k] =  w.pieces_weights_square_related[i][j][flip(k)] * piece_square_related_wmult + w.pieces_weights_const[i][j] * piece_wmult;
+					pieces_weights[i][(j << 1) | 1][k] = -w.pieces_weights_square_related[i][j][k      ] * piece_square_related_wmult - w.pieces_weights_const[i][j] * piece_wmult;
 				}
 			}
 		}
 
-		/* Mobility weights
-			They are dynamic and thus are more general, however they will overlap with static weights 
-			Maybe it's good to test different coefficient values for them or even tweak as standalone weights */
-		let mobility_weights_const = [
-			// knight
-			[
-				-1.7 , -1.0 , -0.3 , -0.07,  0.0 ,  0.2 ,  0.35,  0.4 ,
-				 0.4 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,
-				 0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,
-				 0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0
-			],
-			// bishop
-			[
-				-1.7 , -1.2 , -0.5 , -0.3 ,  0.0 ,  0.21,  0.36,  0.45,
-				 0.5 ,  0.5 ,  0.5 ,  0.5 ,  0.5 ,  0.5 ,  0.0 ,  0.0 ,
-				 0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,
-				 0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0
-			],
-			// rook
-			[
-				-1.7 , -1.2 , -1.04, -0.3 , -0.15,  0.0 ,  0.15,  0.25,
-				 0.35,  0.45,  0.5 ,  0.55,  0.6 ,  0.6 ,  0.6 ,  0.0 ,
-				 0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,
-				 0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0
-			],
-			// queen
-			[
-				-2.1 , -1.35, -0.65, -0.3 , -0.2 ,  0.0 ,  0.15,  0.3 ,
-				 0.4 ,  0.5 ,  0.6 ,  0.7 ,  0.8 ,  0.85,  0.9 ,  0.95,
-				 1.0 ,  1.0 ,  1.0 ,  1.0 ,  1.0 ,  1.0 ,  1.0 ,  1.0 ,
-				 1.0 ,  1.0 ,  1.0 ,  1.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0
-			],
-			// king (opening/mittelspiel)
-			[
-				 0.0 ,  0.05,  0.1 ,  0.1 , -0.6 , -0.6 , -0.6 , -0.6 ,
-				-0.6 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,
-				 0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,
-				 0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0
-			],
-			// king (endspiel)
-			[
-				-1.31, -0.5 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,
-				 0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,
-				 0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,
-				 0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0 ,  0.0
-			]
-		];
 		// transform and apply coefficients
 		let mut mobility_weights = [[[0.0; 32]; 12]; 2];
 		for j in 0..5 {
 			for k in 0..32 {
 				for i in 0..2 {
-					mobility_weights[i][ (j + 1) << 1     ][k] =  mobility_weights_const[j][k] * mobility_wmult;
-					mobility_weights[i][((j + 1) << 1) | 1][k] = -mobility_weights_const[j][k] * mobility_wmult;
+					mobility_weights[i][ (j + 1) << 1     ][k] =  w.mobility_weights_const[j][k] * mobility_wmult;
+					mobility_weights[i][((j + 1) << 1) | 1][k] = -w.mobility_weights_const[j][k] * mobility_wmult;
 				}
 			}
 		}
 		for k in 0..32 {
-			mobility_weights[1][10][k] =  mobility_weights_const[5][k] * mobility_wmult;
-			mobility_weights[1][11][k] = -mobility_weights_const[5][k] * mobility_wmult;
+			mobility_weights[1][10][k] =  w.mobility_weights_const[5][k] * mobility_wmult;
+			mobility_weights[1][11][k] = -w.mobility_weights_const[5][k] * mobility_wmult;
 		}
 
-		/* More dynamic weights */
-		let mut turn_weights_pre	=  [ 1.15,  0.1 ];
-		let mut align_weights_pre   = [[ 0.35,  0.3 ,  0.3 ,  0.21,  0.15,  0.3 ], [ 0.2 ,  0.2 ,  0.1 ,  0.05,  0.1,  0.1 ]];
-		let mut battery_weights_pre =  [ 0.21,  0.22,  0.45,  0.11];
-		let mut dp_weight			=    0.75;
-		let mut pp_weights			=  [ 1.21,  1.11,  1.42];
-		let mut outpost_weight_pre  =  [ 0.4 ,  0.2 ];
-		let mut dan_possible_pre    =  [ 0.4 ,  0.6 ,  0.3 ];
-
 		/* Apply coefficients */
-		turn_weights_pre.iter_mut().for_each(|w| *w = *w * mobility_wmult);
-		align_weights_pre.iter_mut().for_each(|arr| arr.iter_mut().for_each(|w| *w = *w * align_wmult));
-		battery_weights_pre.iter_mut().for_each(|w| *w = *w * battery_wmult);
-		dp_weight += -0.25 + (-pawn_structure_wmult).exp2() * 0.5;
-		pp_weights.iter_mut().for_each(|w| *w = *w * pawn_structure_wmult);
-		outpost_weight_pre.iter_mut().for_each(|w| *w = *w * minor_piece_pos_wmult);
-		dan_possible_pre.iter_mut().for_each(|w| *w = *w * minor_piece_pos_wmult);
+		w.turn_weights_pre.iter_mut().for_each(|w| *w = *w * mobility_wmult);
+		w.align_weights_pre.iter_mut().for_each(|arr| arr.iter_mut().for_each(|w| *w = *w * align_wmult));
+		w.battery_weights_pre.iter_mut().for_each(|w| *w = *w * battery_wmult);
+		w.dp_weight += -0.25 + (-pawn_structure_wmult).exp2() * 0.5;
+		w.pp_weights.iter_mut().for_each(|w| *w = *w * pawn_structure_wmult);
+		w.outpost_weight_pre.iter_mut().for_each(|w| *w = *w * minor_piece_pos_wmult);
+		w.dan_possible_pre.iter_mut().for_each(|w| *w = *w * minor_piece_pos_wmult);
 
 		/* Transform */
-		let mut turn_weights 	= [turn_weights_pre.clone(); 2];
+		let mut turn_weights 	= [w.turn_weights_pre.clone(); 2];
 		turn_weights[1][1] = -turn_weights[0][1];
 		turn_weights[1][0] = 1.0 / turn_weights[0][0];
-		let mut battery_weights = [battery_weights_pre.clone(); 2];
+		let mut battery_weights = [w.battery_weights_pre.clone(); 2];
 		battery_weights[1].iter_mut().for_each(|w| *w = *w * -1.0);
-		let mut outpost_weights	= [outpost_weight_pre.clone(); 2];
+		let mut outpost_weights	= [w.outpost_weight_pre.clone(); 2];
 		outpost_weights[1].iter_mut().for_each(|w| *w = *w * -1.0);
-		let mut dan_possible	= [dan_possible_pre.clone(); 2];
+		let mut dan_possible	= [w.dan_possible_pre.clone(); 2];
 		dan_possible[1].iter_mut().for_each(|w| *w = *w * -1.0);
 		let mut align_weights 	= [[[0.0; 6]; 2]; 2];
 		for i in 0..2 {
 			for k in 0..6 {
-				align_weights[i][0][k] =  align_weights_pre[i][k];
-				align_weights[i][1][k] = -align_weights_pre[i][k];
+				align_weights[i][0][k] =  w.align_weights_pre[i][k];
+				align_weights[i][1][k] = -w.align_weights_pre[i][k];
 			}
 		}
 
+		let zobrist = Zobrist::default();
 		let mut cache_perm_vec = Vec::with_capacity(300);
 		cache_perm_vec.push(zobrist.cache_new(board));
 
@@ -343,8 +134,8 @@ impl Chara {
 			turn_weights,
 			align_weights,
 			battery_weights,
-			dp_weight,
-			pp_weights,
+			dp_weight:		w.dp_weight.clone(),
+			pp_weights:		w.pp_weights.clone(),
 			outpost_weights,
 			dan_possible,
 			random_range,
